# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Swagger documentation",
        "version": "1.0"
    },
    "paths": {
...
        "/api/user-settings/{uid}": {
            "get": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Get settings list for uid",
                "operationId": "c649b5e4daf1f565d281532a23cc60b0",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "uid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Settings response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Settings"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Adds a new Settings",
                "operationId": "659174abf6e543b55dc6594832568733",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value"
                                ],
                                "properties": {
                                    "uid": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string"
                                    },
                                    "value": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "lr response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Settings"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user-settings/{uid}/{sid}": {
            "get": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Get Settings by {sid}",
                "operationId": "455dfc9e523170d0df5c706ba6053ddc",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "uid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "sid",
                        "in": "path",
                        "description": "cid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lr response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Settings"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Settings"
                ],
                "description": "deletes a single setting by sid",
                "operationId": "eb7830af24317644af7ac58c400f99d4",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "uid",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sid",
                        "in": "path",
                        "description": "sid to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "setting deleted"
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                }
            }
        },
        "/api/user-settings/{uid}/search": {
            "get": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Get Settings for user {uid} and optional {type}",
                "operationId": "5de9041501e7ea8fd374e1972130e21e",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "uid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "show"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "lr response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Settings"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/user-settings/{uid}/update": {
            "post": {
                "tags": [
                    "User Settings"
                ],
                "summary": "Updates settings for uid and type",
                "operationId": "5c834cc6a080e80b8ad51a8b367c49e5",
                "parameters": [
                    {
                        "name": "uid",
                        "in": "path",
                        "description": "uid",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "type",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "show"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "value"
                                ],
                                "properties": {
                                    "value": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": {
                                            "column_pos": 0
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Settings": {
                "type": "object",
                "allOf": [
                    {
                        "required": [
                            "uid"
                        ],
                        "properties": {
                            "sid": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "uid": {
                                "type": "integer",
                                "format": "int64"
                            },
                            "type": {
                                "type": "string"
                            },
                            "value": {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "skip_already": 0,
                                        "use_yandex_xml": 0,
                                        "select_gid": 0,
                                        "f_use_base": 1,
                                        "f_skip_already": 1,
                                        "f_check_old": 1
                                    }
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer {token})",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}